Corrected Linear Search and Linear Regression Line

+++ src/com/kevinthelago/Algorithm.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.concurrent.Callable;
 
 public interface Algorithm {
-    List<Long> run(int n);
+    List<Integer> run(int n);
 
-    Scene createScene(List<Long> times);
+    Scene createScene(List<Integer> times);
 }
